Move graphics conversion helpers
