+++ Tests/OpenSwiftUICoreTests/Graphic/GraphicsConversionsTests.swift
@@ -1,14 +1,14 @@
 //
-//  CAFrameRateRangeUtilTests.swift
+//  GraphicsConversionsTests.swift
 //  OpenSwiftUICoreTests
 
-#if (os(iOS) || os(visionOS)) && canImport(QuartzCore)
 import Testing
 @_spi(ForOpenSwiftUIOnly)
 import OpenSwiftUICore
-import QuartzCore
+import OpenQuartzCoreShims
 
-struct CAFrameRateRangeUtilTests {
+struct GraphicsConversionsTests {
+    #if (os(iOS) || os(visionOS)) && canImport(QuartzCore)
     @Test(arguments: [
         (.zero, .default),
         (0.05, .init(minimum: 20, maximum: 60, preferred: 20)), // 20
@@ -16,9 +16,9 @@ struct CAFrameRateRangeUtilTests {
         (0.02, .default), // 50
         (0.0125, .init(minimum: 80, maximum: 80, preferred: 80)), // 80
     ] as [(Double, CAFrameRateRange)])
-    func initWithInterval(interval: Double, expected: CAFrameRateRange) {
+    func initCAFrameRateRangeWithInterval(interval: Double, expected: CAFrameRateRange) {
         let range = CAFrameRateRange(interval: interval)
         #expect(range == expected)
     }
+    #endif
 }
-#endif
